10% improvement in RPS

Issue (2 ClearFast lines)
![cpu-trace](http://aoa.blob.core.windows.net/aspnet/ClearFast.png)

Before
> Running 30s test @ http://..../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     9.46ms   18.88ms 428.30ms   96.09%
    Req/Sec    46.14k     9.03k  194.88k    92.81%
  44125265 requests in 30.10s, 5.42GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 1465986.87
Transfer/sec:    184.55MB

After
> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     9.76ms   20.82ms 366.01ms   95.03%
    Req/Sec    52.65k     5.22k  116.78k    90.51%
  48788229 requests in 30.10s, 6.00GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 1620849.91
Transfer/sec:    204.04MB


Resolves #492