@lodejard 2015-12-17
Hmm... If the headers aren't zeroed out after a request, and the application is responding with computed strings, couldn't this result in more computed string entering the gen1/gen2 heaps?

I mean, yes better for raw RPS, but worse for normal application loads?