@benaadams 2015-12-17
>  result in more computed string entering the gen1/gen2 heaps?

Assuming the setters are clearing the references (either by setting a full StringValue set, or by setting the non-set components to null/empty) then it would only be rarely set headers that would be promoted, as the other ones would be cleared in the next request set - its just avoiding both a double clear (clear+set); and a clear of items that were never set in the first place.

Not sure about the implications pre-encoded change though; haven't looked too closely.