@lodejard 2015-12-17
I'm not so sure that's a rare case - typically browsers will keep a couple connections open at a time, and you requests to arrive in small bursts with longish delays between. If reset doesn't zero the memory then all of the strings from the previous response will still be referenced by the header object during those long delays. So if you have a thousand browsers in the world with your web page open, and the GC kicks off, all of the strings from the most recent response for the thousand connections will move into the next generation.

Pre-encoded and string literals are no problem though - those should be treated like static readonly constants once they are created, and should move into the gen2 heap and stay there forever.