@benaadams 2015-12-18
Solution for that might be to pool the header collections between connections so they aren't active in the keep-alive wait state?